Un-iframe the Google results
