+++ lib/flows/google-flow.js
@@ -6,6 +6,7 @@ import * as mediaHelper from '../media-helper.js';
 import * as driverManager from '../driver-manager.js';
 import * as dataHelper from '../data-helper.js';
 
+import GoogleAdPreviewTool from '../pages/external/google-ad-preview-tool.js';
 import GoogleSearchPage from '../pages/external/google-search.js';
 
 export default class GoogleFlow {
@@ -20,17 +21,19 @@ export default class GoogleFlow {
 		}
 	}
 
-	search( query, culture ) {
+	search( query, data ) {
 		var d = webdriver.promise.defer();
 		var driver = this.driver;
 		var screenSize = this.screenSize;
-		this.GoogleSearchPage = new GoogleSearchPage( driver, { location: 'Sao Paulo', culture: 'pt-br', googleDomain: 'com' } );
+		this.GoogleAdPreviewTool = new GoogleAdPreviewTool( driver, data );
 
 		driver.takeScreenshot().then( function( data ) {
 			// mediaHelper.writeScreenshot( data, culture.toUpperCase() + '_' + screenSize + 'GOOGLE', true );
 		} );
 
-		this.GoogleSearchPage.search( query );
+		this.GoogleAdPreviewTool.search( query );
+		var searchUrl = this.GoogleAdPreviewTool.getSearchPageUrl();
+		this.GoogleSearchPage = new GoogleSearchPage( driver );
 		this.GoogleSearchPage.getAdUrl( 'wordpress.com' );
 
 		d.fulfill( true );
