+++ specs-i18n/search-wordpress-on-google.js
@@ -9,6 +9,7 @@ const mochaTimeOut = config.get( 'mochaTimeoutMS' );
 const startBrowserTimeoutMS = config.get( 'startBrowserTimeoutMS' );
 
 const culture = process.env.CULTURE || 'en';
+const location = process.env.LOCATION || 'New York City';
 
 var driver;
 
@@ -27,6 +28,6 @@ test.describe( 'Search for WordPress on Google', function() {
 
 	test.it( `Can load Google`, function() {
 		const googleFlow = new GoogleFlow( driver, 'desktop' );
-		googleFlow.search( 'wordpress', culture );
+		googleFlow.search( 'wordpress', { location, culture, googleDomain: 'com' }  );
 	} );
 } );
